requirements-seo.txt
google-generativeai>=0.8.0
python-frontmatter>=1.1.0
